#1693 move applyLocalTimeZoneOffset to dateTimeUtils
