+++ wls-gui-wahllokalsystem/src/composables/common/dateTimeFormatter.ts
@@ -29,15 +29,6 @@ export function useDateTimeFormatter() {
     return `${hour}${TIME_FIELD_SEPARATOR}${minute}`;
   };
 
-  const applyLocalTimezoneOffset = function (date: Date | string): Date {
-    const mappedUhrzeit = new Date(date);
-    mappedUhrzeit.setHours(
-      mappedUhrzeit.getHours() -
-        Math.trunc(mappedUhrzeit.getTimezoneOffset() / 60)
-    );
-    return mappedUhrzeit;
-  };
-
   const getDateFromTimeString = function (timeString: string): Date {
     // Validate time string format (HH:MM)
     if (!timeString || !/^\d{1,2}:\d{1,2}(?::\d{1,2})?$/.test(timeString)) {
@@ -115,7 +106,6 @@ export function useDateTimeFormatter() {
     toHhMmSs,
     toHhMm,
     toYyyyMmDd,
-    applyLocalTimezoneOffset,
     getDateFromTimeString,
     toGermanDate,
     toGermanDateWithLongMonth,
