+++ wls-gui-wahllokalsystem/src/composables/common/dateTimeUtils.ts
@@ -3,7 +3,17 @@ export function useDateTimeUtils() {
     return !isNaN(date.getTime());
   }
 
+  const applyLocalTimezoneOffset = function (date: Date | string): Date {
+    const mappedUhrzeit = new Date(date);
+    mappedUhrzeit.setHours(
+      mappedUhrzeit.getHours() -
+        Math.trunc(mappedUhrzeit.getTimezoneOffset() / 60)
+    );
+    return mappedUhrzeit;
+  };
+
   return {
     isValidDate,
+    applyLocalTimezoneOffset,
   };
 }
